First draft of typescript screeps
Just the real basics.  Started a wrapper and declaration of official API.

This will automatically create a dud harvester but wrapper will provide a close experience to the original API
with added utility functions.

Main is dead simple
   const game = new GameWrap();

    console.log(`
===========  ${game.time} =
    * Total number of spawns ${game.spawns.count}
    * Total number of creeps ${game.creeps.count}
    `);

    if(game.creeps.count <= 0) {
        game.spawns.central.spawnCreep([WORK, MOVE, CARRY], "Harvester1")
    }

which will print this in the console
[10:25:33 PM]
===========  1712 =
    * Total number of spawns 1
    * Total number of creeps 1

However trying this with the original API will yield undefined for the count property.

was fun to find how to use typescript and javascript to bend it to my will :)